+++ Kernel/TTime.h
@@ -49,7 +49,7 @@ extern const TTime Time;
 class TimeWheel
 {
 public:
-	uint	Expire;		// 到期时间，毫秒
+	UInt64	Expire;		// 到期时间，毫秒
 	ushort	Sleep;		// 睡眠时间，默认0毫秒
 
 	TimeWheel(uint ms);
